requirements.txt
bs4
python-twitter
